feat: useRegulationsStore separated into two files
